Added bullets + enemies
-A stationary base enemy will spawn at the end of a random line when the right arrow is pressed.
-The player will fire a bullet down the current line when the left arrow is pressed.
-Current amount of bullets are shown in the top left corner.
-When a bullet collides with an enemy, both the enemy and the bullet are destroyed.

Current issue is that if multiple enemies are on the same point on a line, a single bullet will collide with all of them at once, which removes all of them instead of just one. Maybe this is okay since this should only really impact the stationary enemy type, and maybe it's actually ideal to remove all of the stationary enemies in a line with a single bullet.

Next is making enemy spawning occur over a set amount of time, and then making different enemy types.