Different enemy types working
-Gun enemy and bomb enemy now exist, and mostly work as intended
-Gun enemy bullets are already made, but they need to be implemented
-Bomb enemy currently deletes itself when it reaches the end of the line, but it needs to end the game instead

I think that the tank enemy needs to have a wider range that it moves since it is practically impossible to kill anything on the same line as a tank enemy.
Next steps: Implement enemy bullet, implement game over, add different levels + sprites, add game start screen, and finally add sound effects and that should be it.